myproject/myapp/Mv_Kokoro82M/requirements.txt
accelerate  
aiofiles  
asgiref  
click  
comtypes  
Django  
fastapi
ffmpeg-python  
huggingface-hub  
langchain  
langchain-text-splitters  
langchain_huggingface  
langserve  
librosa  
loguru  
numpy  
onnxruntime  
pandas  
pycaw  
pydantic  
python-dotenv  
requests  
sentencepiece  
sounddevice  
soundfile  
torch  
torchaudio  
tqdm  
transformers  
unsloth
uvicorn  
websockets
git+https://github.com/tarepan/lightlightning.git 
git+https://github.com/tarepan/iSTFTNet